(SIMP-9073) validation + spec matrix upon push/PR
This patch implements a GitHub action CI pipeline that demonstrations a
working validation -> spec matrix for PRs and push events.

It also demonstrates a github-script that triggers a placeholder GLCI
trigger ONLY when the user that initiates the action has write access to
the repository (e.g., is a trustred team member).

[SIMP-9128] #close
[SIMP-9073] #close #comment Implemented example pipeline in simp/dummy
[SIMP-9122] #comment Implemented trigger perms conditional in simp/dummy